api/tests/tests_users/base/test_base_user_views.py
from django.test import TestCase
from rest_framework.test import APIClient
from UserProfile.models import Customer
from core.models import User


class TestBaseUserViews(TestCase):
    def setUp(self):
        self.client = APIClient()
        
        return super().setUp()