requirements.txt
asgiref==3.8.1
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-watch==4.2.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
watchdog==5.0.3
